Create SAMP_ReadLastLine.ahk
A SAMP chatlog file is read like this:
(
Line 1

Line 2

Line 3


)

Not including the bracket lines. Basically it has a line, the. A blank line, then a line, etc. And the bottom is a blank line, so once you get to that line, you need to minus two from the total number of lines to get the latest line.

To use this, use it like this:
OutputVar:=SAMP_ReadLast(FlieIn)

Output var will be the last line in your chatlog.